+++ codejam/client/rootwidget.kv
@@ -1,0 +1,8 @@
+RootWidget:
+    ws: False
+    Screen:
+        Button:
+            on_release: root.current = 'whiteboard'
+    WhiteBoardScreen:
+        id: whiteboard
+        name: 'whiteboard'
